Move import machinery out of __init__.py
